+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## [0.1.3](https://github.com/martinsjavacode/finn/compare/v0.1.2...v0.1.3) (2026-06-15)
+
+
+### Features
+
+* **rls:** add cards RLS permission function and property-based tests ([853fd74](https://github.com/martinsjavacode/finn/commit/853fd746ee2a22480f4068f9c1eaa4630a3f7c8b))
+* **transactions:** add batch selection and bulk payment actions ([f4485cc](https://github.com/martinsjavacode/finn/commit/f4485cc2400a1aeee7807b9fa697a253e330d071))
+
 ## [0.1.2](https://github.com/martinsjavacode/finn/compare/v0.1.1...v0.1.2) (2026-05-31)